dockerfiles/requirements.txt
aiodns==2.0.0
bidict==0.21.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
CouchDB==1.2
Flask==1.1.2
Flask-SocketIO==5.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
paho-mqtt==1.5.1
pandas==1.1.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
requests==2.25.1
six==1.15.0
slixmpp==1.7.0
typing==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
